+++ lib/src/configuration/better_player_controls_configuration.dart
@@ -78,8 +78,8 @@ class BetterPlayerControlsConfiguration {
   final Duration controlsHideTime;
 
   ///Parameter used to build custom controls
-  final Widget Function(BetterPlayerController controller)?
-      customControlsBuilder;
+  final Widget Function(BetterPlayerController controller,
+      Function(bool) onPlayerVisibilityChanged)? customControlsBuilder;
 
   ///Parameter used to change theme of the player
   final BetterPlayerTheme? playerTheme;